Refactor game code to work with button clicks instead of prompts
Harder than I thought it would be but I was able to assign the
playerChoice value and run the playRound function on a single button
click. Most of my problems were related to syntax and not realizing
I needed to remove the parameters from playRound since they were being
supplied from elsewhere in the code now.